feat : VideoPlayer.tsx삭제하고 내부의 ControlBar컴포넌트 분리
