+++ demo/src/App.tsx
@@ -6,6 +6,7 @@ import SubtitleOption from './components/SubtitleOption';
 import { LanguageCode } from './interfaces/Scripts';
 import scriptsMockData from './mocks/subtitleMockData';
 import { mockUrl } from './mocks/mockUrl';
+import ControlBar from './components/ControlBar';
 
 type Mode = 'line' | 'block';
 
@@ -31,6 +32,46 @@ function App() {
     setCurrentTime(state.playedSeconds);
   };
 
+  // --- todo : ControlBar에서만 쓰이는 속성 ControlBar로 내부로 이동
+  const [isPlaying, setIsPlaying] = useState(false);
+  const [volume, setVolume] = useState(false);
+
+  const handlePlayPause = () => {
+    setIsPlaying((prev) => !prev);
+  };
+
+  const handleVolumeChange = () => {
+    setVolume((prev) => !prev);
+  };
+  const handleSeekForward = () => {
+    if (playerRef.current) {
+      playerRef.current.seekTo(
+        (playerRef.current.getCurrentTime() || 0) + 10,
+        'seconds',
+      );
+      console.log(playerRef.current.getCurrentTime());
+    }
+  };
+
+  const handleSeekBackward = () => {
+    if (playerRef.current) {
+      playerRef.current.seekTo(
+        (playerRef.current.getCurrentTime() || 0) - 10,
+        'seconds',
+      );
+    }
+  };
+
+  const BasicControlBarProps = {
+    handlePlayPause,
+    handleVolumeChange,
+    handleSeekBackward,
+    handleSeekForward,
+    isPlaying,
+    volume,
+  };
+  // ---
+
   return (
     <div className="App">
       {/* TODO(@smosco): progressInterval을 더 짧게 주자 */}
@@ -39,6 +80,10 @@ function App() {
         url={mockUrl}
         playing
         onProgress={handleProgress}
+      />
+      <ControlBar
+        playerRef={playerRef}
+        BasicControlBarProps={BasicControlBarProps}
       />
 
       <SubtitleOption
